+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [Unreleased]
 - OpenAI Translator now supports Azure OpenAI.
-- OpenAI Translator now supports the GPT-3.5-Turbo-16K model.
 - OpenAI Translator now offers API endpoint configuration.
+- The model list of OpenAI Translator has been updated.
 
 ## [3.5.6] (2023/11/19)